### 2026-01-18 #9849 @editorialbot
[PRE REVIEW]: scphylo-tools: A Python toolkit for single-cell tumor phylogenetic analysis
**Submitting author:** @faridrashidi (<a href="http://orcid.org/0000-0003-4103-4904">Farid Rashidi Mehrabadi</a>)
**Repository:** https://github.com/faridrashidi/scphylo-tools
**Branch with paper.md** (empty if default branch):
**Version:** v0.0.5
