### 2026-02-11 d2cf430
refactor(utils): ♻️ Replace pkg_resources with importlib.resources
