### 2026-02-11 d644f99
style(_utils): 💄 Improve formatting and consistency in logging
