### 2026-03-23 c8e6020
📝 Update affiliation to full name in paper metadata
