### 2026-03-23 #9849 @willgearty
@faridrashidi could you please suggest 5 reviewers from our [reviewer database](https://reviewers.joss.theoj.org/lookup) or your own ([non-conflicted](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy)) extended network? Their GitHub handles are most useful to receive but please don't use "@" to reference them since it will prematurely ping them.

(I recognize there are two volunteers, but it would be good to have a broader set of people to pull from)
