### 2026-03-23 #9849 @faridrashidi
Hi @willgearty, I’ve identified a few people who seem to have deep expertise in cancer genomics, single-cell omics, and bioinformatics:
- skanwal
- rachelicr
- SamPIngram
- rcannood
- dolchan
