### 2026-05-09 c8dba16
🔧 Update for improved dependency management
