### 2026-06-28 faridrashidi/scphylo-tools#13 @rachelicr
Documentation mismatch for installation
openjournals/joss-reviews#10589

There is an error in the installation instructions here:
https://scphylo-tools.readthedocs.io/en/latest/getting_started.html#Installation

It is correct here:
https://scphylo-tools.readthedocs.io/en/latest/installation.html#using-pypi

Additional note on installation - my original clean installation into conda with python 3.12 used a later version of setup tools that didn't include pkg_resources, I had to downgrade: `pip install "setuptools<81"`
