### 2026-06-28 bc95cee
🔥 Remove setuptools dependency from project
