### 2026-06-28 70821f2
🎨 Improve code structure and formatting
