+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "scphylo-tools"
-dynamic = ["version"]
+version = "0.0.5"
 description = "A python toolkit for single-cell tumor phylogenetic analysis"
 requires-python = ">=3.11"
 dependencies = [
@@ -66,7 +66,13 @@ Download = "https://github.com/faridrashidi/scphylo-tools"
 scphylo = "scphylo.commands._cli:cli"
 
 [project.optional-dependencies]
-dev = ["pre-commit", "pytest-cov", "pytest-xdist", "bumpversion", "ipykernel"]
+dev = [
+    "pre-commit",
+    "pytest-cov",
+    "pytest-xdist",
+    "bump-my-version",
+    "ipykernel",
+]
 docs = [
     "furo",
     "ipython",
@@ -97,3 +103,23 @@ where = ["src"]
 
 [tool.cibuildwheel]
 build = "cp3*-* pp3*-*"
+
+[tool.bumpversion]
+current_version = "0.0.5"
+commit = true
+message = "🔖 Release {new_version}"
+tag = true
+tag_name = "v{new_version}"
+parse = '(?P<major>\d+)\.(?P<minor>\d+)\.(?P<patch>\d+)'
+serialize = ["{major}.{minor}.{patch}"]
+
+[[tool.bumpversion.files]]
+filename = "src/scphylo/__init__.py"
+search = '__version__ = "{current_version}"'
+replace = '__version__ = "{new_version}"'
+
+[[tool.bumpversion.files]]
+filename = "uv.lock"
+regex = true
+search = '(?ms)(\[\[package\]\]\nname = "scphylo-tools"\nversion = )"{current_version}"'
+replace = '\1"{new_version}"'
