+++ .bumpversion.cfg
@@ -1,10 +1,0 @@
-[bumpversion]
-current_version = 0.0.5
-commit = True
-message = chore(release): 🔧 Bump version {current_version} → {new_version}
-tag = True
-tag_name = v{new_version}
-files = setup.py src/scphylo/__init__.py
-parse = (?P<major>\d+)\.(?P<minor>\d+)\.(?P<patch>\d+)
-serialize = 
-	{major}.{minor}.{patch}
