+++ Makefile
@@ -1,8 +1,14 @@
+VALID_RELEASE_PARTS := patch minor major
+RELEASE_ARGS = $(filter-out $@,$(MAKECMDGOALS))
+
+.PHONY: help clean lint test doc build install install-dev cythonize release $(VALID_RELEASE_PARTS)
+
 help:
 	@echo "available commands"
 	@echo " - lint         : run linting and flaking"
 	@echo " - test         : run all unit tests"
 	@echo " - doc          : build the documentation"
+	@echo " - release      : bump version and create a release tag"
 	@echo " - install      : install the package"
 	@echo " - install-dev  : install the package in dev mode"
 	@echo " - cythonize    : cythonize all pyx files"
@@ -46,3 +52,13 @@ install-dev:
 
 cythonize:
 	CYTHONIZE=1 pip install -e .
+
+release:
+	@if [ "$(words $(RELEASE_ARGS))" -ne 1 ] || [ -n "$(filter-out $(VALID_RELEASE_PARTS),$(RELEASE_ARGS))" ]; then \
+		echo "usage: make release [patch|minor|major]"; \
+		exit 1; \
+	fi
+	uv run --extra dev bump-my-version bump $(RELEASE_ARGS)
+
+$(VALID_RELEASE_PARTS):
+	@:
