### 2026-06-28 faridrashidi/scphylo-tools#13 @faridrashidi
@rachelicr Thanks for reporting this.

The getting started docs have been updated to use the correct PyPI package name.

I also checked the pkg_resources / setuptools issue. Instead of requiring users to downgrade setuptools, I removed setuptools from the runtime dependencies and kept it only as a build-system dependency. This allows installation with latest setuptools versions where pkg_resources is no longer available.
