### 2026-07-13 81879cb
🔧 Update data handling in functions
