### 2026-07-18 f941fac
🔧 Update Makefile for improved commands
