### 2026-07-18 ac30a9a
✨ Update paper for clarity
