### 2026-07-18 659bef1
ci: install uv before Make targets
