### 2026-07-18 087bd17
✨ Move from uv to pixi
