### 2026-07-18 8724634
🔥 Remove unused configuration files
