+++ src/scphylo/ul/_utils.py
@@ -15,6 +15,7 @@ import scphylo as scp
 
 
 def log_input(df_in):
+    """Log dimensions and genotype-state counts for an input matrix."""
     size = df_in.shape[0] * df_in.shape[1]
     scp.logg.info(f"input -- size: {df_in.shape[0]}x{df_in.shape[1]}")
     scp.logg.info(
@@ -33,6 +34,7 @@ def log_input(df_in):
 
 
 def log_output(df_out, running_time):
+    """Log dimensions, genotype states, validity, and runtime for an output."""
     size = df_out.shape[0] * df_out.shape[1]
     scp.logg.info(f"output -- size: {df_out.shape[0]}x{df_out.shape[1]}")
     scp.logg.info(
@@ -60,6 +62,7 @@ def log_output(df_out, running_time):
 
 
 def log_flip(df_in, df_out):
+    """Log genotype flips and inferred error rates between two matrices."""
     flips_0_1, flips_1_0, flips_na_0, flips_na_1 = count_flips(
         df_in.values, df_out.values, 3
     )
@@ -74,6 +77,7 @@ def log_flip(df_in, df_out):
 
 
 def calc_nll_matrix(df_in, df_out, alpha, beta):
+    """Calculate the negative log-likelihood of an inferred genotype matrix."""
     if alpha == 0 or beta == 0:
         return None
     columns = np.intersect1d(df_in.columns, df_out.columns)
@@ -102,6 +106,7 @@ def calc_nll_matrix(df_in, df_out, alpha, beta):
 
 
 def stat(df_in, df_out, alpha, beta, running_time):
+    """Log a summary comparing input and inferred genotype matrices."""
     log_input(df_in)
     log_output(df_out, running_time)
     log_flip(df_in, df_out)
@@ -110,6 +115,8 @@ def stat(df_in, df_out, alpha, beta, running_time):
 
 
 def parse_params_file(filename):
+    """Parse simulation and error-rate parameters encoded in a filename."""
+
     def _parse_params_file_helper(param):
         try:
             value = basename.split(f"{param}_")[1]
@@ -144,6 +151,7 @@ def parse_params_file(filename):
 
 
 def parse_log_file(filename):
+    """Parse runtime and error-rate statistics from a solver log file."""
     result = {}
     _, basename = dir_base(filename)
     result["tool"] = basename.split(".")[-1]
@@ -164,6 +172,7 @@ def parse_log_file(filename):
 
 
 def parse_score_file(filename):
+    """Parse key-value scores from a solver score file."""
     result = {}
     _, basename = dir_base(filename)
     result["tool"] = basename.split(".")[-1]
@@ -175,6 +184,7 @@ def parse_score_file(filename):
 
 
 def count_flips(I_mtr, O_mtr, na_value=3):
+    """Count state changes between input and output genotype matrices."""
     flips_0_1 = 0
     flips_1_0 = 0
     flips_na_0 = 0
@@ -194,6 +204,7 @@ def count_flips(I_mtr, O_mtr, na_value=3):
 
 
 def infer_rates(I_mtr, O_mtr, na_value=3):
+    """Infer false-negative, false-positive, and missing-data rates."""
     flips_0_1, flips_1_0, flips_na_0, flips_na_1 = count_flips(I_mtr, O_mtr, na_value)
     fn_rate = flips_0_1 / ((O_mtr == 1) & (I_mtr != na_value)).sum()
     fp_rate = flips_1_0 / ((O_mtr == 0) & (I_mtr != na_value)).sum()
@@ -303,43 +314,52 @@ def is_conflict_free_gusfield(df_in):
 
 
[…]
