+++ tests/test_logging.py
@@ -1,10 +1,15 @@
+"""Verify package logging behavior."""
+
 import pytest
 
 import scphylo as scp
 
 
 class TestLogging:
+    """Exercise all public logging levels."""
+
     def test_logging(self):
+        """Verify normal messages and error escalation across log levels."""
         scp.logg.print_version()
         scp.logg.debug("DEBUG")
         scp.logg.hint("HINT")
