+++ examples/comparison/compute_gs.py
@@ -1,9 +1,7 @@
-"""
-Comparing two trees by Genotype-Similarity
-------------------------------------------
+"""Genotype-similarity comparison.""" "\n-------------------------------"
 
-This example shows how to compare/measure two inferred genotype data (trees).
-"""
+# %%
+# This example shows how to compare or measure two inferred genotype trees.
 
 import scphylo as scp
 
