+++ examples/comparison/compute_dl.py
@@ -1,9 +1,7 @@
-"""
-Comparing two trees by Different-Lineage
-----------------------------------------
+"""Different-lineage comparison.""" "\n-----------------------------"
 
-This example shows how to compare/measure two inferred genotype data (trees).
-"""
+# %%
+# This example shows how to compare or measure two inferred genotype trees.
 
 import scphylo as scp
 
