+++ examples/reconstruction/compute_siclonefit.py
@@ -1,10 +1,8 @@
-"""
-Reconstruct tree by SiCloneFit
-------------------------------
+"""SiCloneFit reconstruction.""" "\n--------------------------"
 
-This example shows how to construct a phylogenetic tree using SiCloneFit on a
-binary single-cell genotype matrix.
-"""
+# %%
+# This example shows how to construct a phylogenetic tree using SiCloneFit on a
+# binary single-cell genotype matrix.
 
 import scphylo as scp
 

[…]
