### 2026-07-18 9f14359
🎨 Improve code formatting and structure
