### 2026-07-18 fd5d802
🐛 Fix disc function mutation handling
