### 2026-07-18 b0b926e
✨ Introduce external tool resolution
