+++ tests/test_tl_solvers_tmp.py
@@ -1,5 +1,10 @@
 """Exercise experimental and optional solver integrations."""
 
+import subprocess
+from pathlib import Path
+
+import pandas as pd
+
 import scphylo as scp
 
 from ._helpers import skip_rpy2, skip_slow
@@ -35,9 +40,88 @@ class TestSolversTmp:
         is_cf = scp.ul.is_conflict_free_gusfield(df_out)
         assert is_cf
 
-    def test_siclonefit(self):
-        """Reserve coverage for the optional SiCloneFit integration."""
-        assert True
+    def test_siclonefit(self, tmp_path, monkeypatch):
+        """Verify the SiCloneFit adapter with paths containing spaces."""
+        df_in = scp.datasets.test()
+        expected = df_in.copy()
+        expected.iloc[:, :] = 0
+        expected.iloc[0, 0] = 1
+
+        tools_dir = tmp_path / "tools with spaces"
+        tools_dir.mkdir()
+        jar_path = tools_dir / "SiCloneFiTComplete.jar"
+        jar_path.write_text("fixture")
+        java_executable = tools_dir / "java"
+        java_executable.write_text("#!/bin/sh\nexit 0\n")
+        java_executable.chmod(0o755)
+
+        def fake_run(args, appname, **kwargs):
+            assert appname == "SiCloneFit"
+            assert args[:3] == [
+                str(java_executable.resolve()),
+                "-jar",
+                str(jar_path.resolve()),
+            ]
+            workdir = Path(args[args.index("-outDir") + 1])
+            best_dir = workdir / "fixture_samples" / "best"
+            best_dir.mkdir(parents=True)
+            output = expected.T.copy()
+            output.index = range(output.shape[0])
+            output.to_csv(
+                best_dir / "best_MAP_predicted_genotype.txt",
+                sep=" ",
+                header=False,
+            )
+            (best_dir / "best_MAP_tree.txt").write_text("(cell1,cell2);\n")
+            return subprocess.CompletedProcess(args, 0)
+
+        monkeypatch.setattr(scp.ul, "run_external", fake_run)
+        df_out, tree = scp.tl.siclonefit(
+            df_in,
+            alpha=0.0000001,
+            beta=0.1,
+            n_restarts=1,
+            n_iters=1,
+            n_burnin=0,
+            return_tree=True,
+            jar_path=jar_path,
+            java_executable=java_executable,
+        )
+
+        pd.testing.assert_frame_equal(df_out, expected)
+        assert tree == "(cell1,cell2);"
+
+    def test_sphyr(self, tmp_path, monkeypatch):
+        """Verify that SPhyR uses checked argv-list execution."""
+        df_in = scp.datasets.test()
+        expected = df_in.copy()
+        expected.iloc[:, :] = 0
+        expected.iloc[0, 0] = 1
+
+        tools_dir = tmp_path / "tools with spaces"
+        tools_dir.mkdir()
+        executable = tools_dir / "sphyr_kDPFC"
+        executable.write_text("#!/bin/sh\nexit 0\n")
+        executable.chmod(0o755)
+
+        def fake_run(args, appname, **kwargs):
+            assert appname == "SPhyR"
+            assert args[0] == str(executable.resolve())
+            output = kwargs["stdout"]
+            output.write("SPhyR output\ncorrected genotype\n")
+            expected.to_csv(output, sep=" ", header=False, index=False)
+            output.flush()
+            return subprocess.CompletedProcess(args, 0)
+
+        monkeypatch.setattr(scp.ul, "run_external", fake_run)
+        df_out = scp.tl.sphyr(
+            df_in,
+            alpha=0.0000001,
+            beta=0.1,
+            executable=executable,
+        )
+
[…]
