### 2026-07-19 4c38fcc
✨ Update colorectal2 dataset and tests
