+++ tests/test_datasets.py
@@ -1,5 +1,6 @@
 """Verify simulation, noise generation, and bundled dataset loaders."""
 
+import numpy as np
 import pytest
 
 import scphylo as scp
@@ -50,6 +51,63 @@ class TestDatasets:
         adata = func()
         assert adata.shape == (n, m)
 
+    def test_colorectal2_variants(self):
+        """Verify the full genotype and read-count variants stay aligned."""
+        genotype = scp.datasets.colorectal2()
+        readcount = scp.datasets.colorectal2(readcount=True)
+
+        assert genotype.shape == readcount.shape == (182, 36)
+        assert genotype.obs_names.equals(readcount.obs_names)
+        assert genotype.var_names.equals(readcount.var_names)
+        np.testing.assert_array_equal(genotype.X, readcount.X)
+        values, counts = np.unique(genotype.X, return_counts=True)
+        assert dict(zip(values, counts, strict=True)) == {0: 5174, 1: 866, 3: 512}
+        assert (genotype.X == 1).any(axis=1).sum() == 86
+
+        assert genotype.obs["phiscs_fig7"].sum() == 78
+        assert genotype.var["phiscs_fig7"].sum() == 25
+        assert genotype.obs["copy_number_profile"].value_counts().to_dict() == {
+            0: 117,
+            1: 32,
+            2: 10,
+            3: 23,
+        }
+
+        phiscs_fig7 = genotype.uns["phiscs_fig7"]
+        assert len(phiscs_fig7["obs_names"]) == 78
+        assert len(phiscs_fig7["var_names"]) == 25
+        assert phiscs_fig7["solution_fig7a"].shape == (78, 25)
+        assert phiscs_fig7["solution_fig7b"].shape == (78, 25)
+
+        assert {"mutant", "total"} <= set(readcount.layers)
+        mutant = readcount.layers["mutant"]
+        total = readcount.layers["total"]
+        assert np.isfinite(mutant).all() and np.isfinite(total).all()
+        assert np.all(mutant >= 0)
+        assert np.all(mutant <= total)
+        assert np.equal(mutant, np.floor(mutant)).all()
+        assert np.equal(total, np.floor(total)).all()
+        assert mutant.sum() == 208955
+        assert total.sum() == 1994705
+        assert np.all(total.sum(axis=1) > 0)
+        assert readcount["MA_94", "TP53_chr17_7577548"].layers["mutant"].item() == 248
+        assert readcount["MA_94", "TP53_chr17_7577548"].layers["total"].item() == 250
+
+        bulk_columns = [column for column in genotype.var if "Count_CO8" in column]
+        assert len(bulk_columns) == 4
+        assert (
+            genotype.var.loc[genotype.var["phiscs_fig7"], bulk_columns]
+            .notna()
+            .all()
+            .all()
+        )
+        assert (
+            genotype.var.loc[~genotype.var["phiscs_fig7"], bulk_columns]
+            .isna()
+            .all()
+            .all()
+        )
+
     def test_load_datasets(self):
         """Verify the expected shapes of all bundled datasets."""
         adata = scp.datasets.example()
@@ -72,7 +130,7 @@ class TestDatasets:
         adata = scp.datasets.colorectal1()
         assert adata.shape == (178, 16)
         adata = scp.datasets.colorectal2()
-        assert adata.shape == (78, 25)
+        assert adata.shape == (182, 36)
         # adata = scp.datasets.colorectal3()
         adata = scp.datasets.erbc()
         assert adata.shape == (47, 40)
