+++ tests/test_datasets.py
@@ -222,6 +222,202 @@ class TestDatasets:
             "gama": 0,
         }
 
+    def test_high_grade_serous_ovarian_cancer3_derivatives(self):
+        """Verify the full HGSOC3 assay and its published subset dimensions."""
+        adata = scp.datasets.high_grade_serous_ovarian_cancer3()
+
+        assert adata.shape == (420, 43)
+        values, counts = np.unique(adata.X, return_counts=True)
+        assert dict(zip(values, counts, strict=True)) == {0: 8099, 1: 4390, 3: 5571}
+        matrix_hash = hashlib.sha256(
+            np.asarray(adata.X, dtype=np.int8).tobytes(order="C")
+        ).hexdigest()
+        assert (
+            matrix_hash
+            == "7e59ef1d8c004f09b6cff293639a2af2dbc523fc400ea413963538ec1341d792"
+        )
+
+        assert adata.obs_names.is_unique
+        assert adata.var_names.is_unique
+        assert adata.obs["sample_id"].value_counts(sort=False).to_dict() == {
+            "left_ovary_site_1": 84,
+            "left_ovary_site_2": 84,
+            "omentum_site_1": 84,
+            "omentum_site_2": 84,
+            "right_ovary_site_1": 84,
+        }
+        obs_payload = (
+            "\n".join(
+                f"{str(index)}\t{str(row.sample_id)}\t{int(row.well_id)}\t"
+                f"{int(row.library)}\t{int(row.total_depth)}\t"
+                f"{int(bool(row.mcpherson_clonal_analysis))}\t"
+                f"{int(bool(row.sciphi_scvilp))}"
+                for index, row in zip(
+                    adata.obs_names, adata.obs.itertuples(), strict=True
+                )
+            )
+            + "\n"
+        )
+        assert (
+            hashlib.sha256(obs_payload.encode()).hexdigest()
+            == "87e4298d0970005bb711fe898ed698cabd844b1d3d1bac1360ae7c1f18366a79"
+        )
+
+        assert {
+            "allele_state",
+            "alt_p_value",
+            "mutant",
+            "ref_p_value",
+            "total",
+        } <= set(adata.layers)
+        allele_state = adata.layers["allele_state"]
+        values, counts = np.unique(allele_state, return_counts=True)
+        assert dict(zip(values, counts, strict=True)) == {
+            0: 8099,
+            1: 3804,
+            2: 586,
+            3: 5571,
+        }
+        state_hash = hashlib.sha256(
+            np.asarray(allele_state, dtype=np.int8).tobytes(order="C")
+        ).hexdigest()
+        assert (
+            state_hash
+            == "fe368c36858bee47973b5f5ea1f360789d93a91af5d73cc995acca4953b5aa7b"
+        )
+        np.testing.assert_array_equal(
+            adata.X,
+            np.where(allele_state == 3, 3, (allele_state > 0).astype(np.int8)),
+        )
+
+        mutant = adata.layers["mutant"]
+        total = adata.layers["total"]
+        called = total >= 50
+        ref_present = adata.layers["ref_p_value"] < 1e-6
+        alt_present = adata.layers["alt_p_value"] < 1e-6
+        expected_state = np.full(adata.shape, 3, dtype=np.int8)
+        expected_state[called & ref_present & ~alt_present] = 0
+        expected_state[called & ref_present & alt_present] = 1
+        expected_state[called & ~ref_present & alt_present] = 2
+        np.testing.assert_array_equal(allele_state, expected_state)
+        assert (
+            hashlib.sha256(
[…]
