+++ tests/test_datasets.py
@@ -301,6 +301,195 @@ class TestDatasets:
             "gama": 0,
         }
 
+    def test_high_grade_serous_ovarian_cancer1_derivatives(self):
+        """Verify the full HGSOC1 assay and the infSCITE Figure S22 view."""
+        adata = scp.datasets.high_grade_serous_ovarian_cancer1()
+
+        assert adata.shape == (588, 43)
+        values, counts = np.unique(adata.X, return_counts=True)
+        assert dict(zip(values, counts, strict=True)) == {
+            0: 10774,
+            1: 4747,
+            3: 9763,
+        }
+        assert (
+            hashlib.sha256(
+                np.asarray(adata.X, dtype=np.int8).tobytes(order="C")
+            ).hexdigest()
+            == "a290d79c2bd1f3885a22643d8f3811bdb04baf65e20d514d4dd93e9ccaf5af31"
+        )
+
+        assert adata.obs_names.is_unique
+        assert adata.var_names.is_unique
+        assert adata.obs["sample_id"].value_counts(sort=False).to_dict() == {
+            "omentum_site_1": 126,
+            "omentum_site_2": 126,
+            "right_ovary_site_1": 84,
+            "right_ovary_site_2": 252,
+        }
+        assert adata.obs["variation"].value_counts(sort=False).to_dict() == {
+            "barcode": 5,
+            "standard": 583,
+        }
+        obs_payload = (
+            "\n".join(
+                f"{str(index)}\t{str(row.sample_id)}\t{int(row.well_id)}\t"
+                f"{int(row.library)}\t{str(row.variation)}\t"
+                f"{int(row.total_depth)}\t"
+                f"{int(bool(row.mcpherson_clonal_analysis))}"
+                for index, row in zip(
+                    adata.obs_names, adata.obs.itertuples(), strict=True
+                )
+            )
+            + "\n"
+        )
+        assert (
+            hashlib.sha256(obs_payload.encode()).hexdigest()
+            == "a7866741c951240ce776ee8163a28289d371b8745cf7b6138d75b51f7d16c982"
+        )
+
+        assert {
+            "allele_state",
+            "alt_p_value",
+            "mutant",
+            "ref_p_value",
+            "total",
+        } <= set(adata.layers)
+        allele_state = adata.layers["allele_state"]
+        values, counts = np.unique(allele_state, return_counts=True)
+        assert dict(zip(values, counts, strict=True)) == {
+            0: 10774,
+            1: 3642,
+            2: 1105,
+            3: 9763,
+        }
+        assert (
+            hashlib.sha256(
+                np.asarray(allele_state, dtype=np.int8).tobytes(order="C")
+            ).hexdigest()
+            == "9f3c7fff6a97e87fc27aa2a8de016a307f5a0593b2dd0792b20ce94125787c41"
+        )
+        np.testing.assert_array_equal(
+            adata.X,
+            np.where(allele_state == 3, 3, (allele_state > 0).astype(np.int8)),
+        )
+
+        mutant = adata.layers["mutant"]
+        total = adata.layers["total"]
+        called = total >= 50
+        ref_present = adata.layers["ref_p_value"] < 1e-6
+        alt_present = adata.layers["alt_p_value"] < 1e-6
+        expected_state = np.full(adata.shape, 3, dtype=np.int8)
+        expected_state[called & ref_present & ~alt_present] = 0
+        expected_state[called & ref_present & alt_present] = 1
+        expected_state[called & ~ref_present & alt_present] = 2
+        np.testing.assert_array_equal(allele_state, expected_state)
[…]
