+++ tests/test_datasets.py
@@ -490,6 +490,224 @@ class TestDatasets:
         assert full_panel["false_positive_rate"] == 0.020
         assert full_panel["false_signal_rate"] == 0.436
 
+    def test_high_grade_serous_ovarian_cancer2_derivatives(self):
+        """Verify the full HGSOC2 assay and the infSCITE Figure S23 view."""
+        adata = scp.datasets.high_grade_serous_ovarian_cancer2()
+
+        assert adata.shape == (672, 84)
+        values, counts = np.unique(adata.X, return_counts=True)
+        assert dict(zip(values, counts, strict=True)) == {
+            0: 18283,
+            1: 10299,
+            3: 27866,
+        }
+        assert (
+            hashlib.sha256(
+                np.asarray(adata.X, dtype=np.int8).tobytes(order="C")
+            ).hexdigest()
+            == "419839ab9e60320ddcb35b0a2ea5ea6a717151d650204fbceb874e3a03d17fe1"
+        )
+
+        assert adata.obs_names.is_unique
+        assert adata.var_names.is_unique
+        assert adata.obs["sample_id"].value_counts(sort=False).to_dict() == {
+            "adnexa_site_1": 168,
+            "omentum_site_1": 168,
+            "right_ovary_site_1": 168,
+            "right_ovary_site_2": 168,
+        }
+        obs_payload = (
+            "\n".join(
+                f"{str(index)}\t{str(row.sample_id)}\t{int(row.well_id)}\t"
+                f"{int(row.library)}\t{int(row.total_depth)}\t"
+                f"{int(bool(row.mcpherson_clonal_analysis))}"
+                for index, row in zip(
+                    adata.obs_names, adata.obs.itertuples(), strict=True
+                )
+            )
+            + "\n"
+        )
+        assert (
+            hashlib.sha256(obs_payload.encode()).hexdigest()
+            == "e0987bd38e75d099ef819e2e8c2aac61614b447f6276adbab7744530f81d064b"
+        )
+
+        assert {
+            "allele_state",
+            "alt_p_value",
+            "mutant",
+            "ref_p_value",
+            "total",
+        } <= set(adata.layers)
+        allele_state = adata.layers["allele_state"]
+        values, counts = np.unique(allele_state, return_counts=True)
+        assert dict(zip(values, counts, strict=True)) == {
+            0: 18283,
+            1: 8533,
+            2: 1766,
+            3: 27866,
+        }
+        assert (
+            hashlib.sha256(
+                np.asarray(allele_state, dtype=np.int8).tobytes(order="C")
+            ).hexdigest()
+            == "5f186d0526743635244259b4a49d5e4e5450ea816a1fb0fe3fa6984d6215d011"
+        )
+        np.testing.assert_array_equal(
+            adata.X,
+            np.where(allele_state == 3, 3, (allele_state > 0).astype(np.int8)),
+        )
+
+        mutant = adata.layers["mutant"]
+        total = adata.layers["total"]
+        called = total >= 50
+        ref_present = adata.layers["ref_p_value"] < 1e-6
+        alt_present = adata.layers["alt_p_value"] < 1e-6
+        expected_state = np.full(adata.shape, 3, dtype=np.int8)
+        expected_state[called & ref_present & ~alt_present] = 0
+        expected_state[called & ref_present & alt_present] = 1
+        expected_state[called & ~ref_present & alt_present] = 2
+        np.testing.assert_array_equal(allele_state, expected_state)
+        assert (
+            hashlib.sha256(
[…]
