+++ src/scphylo/datasets/__init__.py
@@ -18,6 +18,7 @@ from scphylo.datasets._datasets import (
     high_grade_serous_ovarian_cancer2,
     high_grade_serous_ovarian_cancer3,
     high_grade_serous_ovarian_cancer_3celllines,
+    isogenic_fibroblast_cell_line,
     melanoma20,
     muscle_invasive_bladder,
     myeloproliferative_neoplasms18,
@@ -48,6 +49,7 @@ __all__ = (
     high_grade_serous_ovarian_cancer2,
     high_grade_serous_ovarian_cancer3,
     high_grade_serous_ovarian_cancer_3celllines,
+    isogenic_fibroblast_cell_line,
     melanoma20,
     muscle_invasive_bladder,
     myeloproliferative_neoplasms18,
