+++ tests/test_datasets.py
@@ -904,6 +904,153 @@ class TestDatasets:
             == "3a750b7fcaedd6587084a54c653221d1a6ecb1c682dcd953d98b439b7d832847"
         )
 
+    def test_isogenic_fibroblast_figure_s9(self):
+        """Verify the lossless SCIPhI Figure S9 fibroblast extraction."""
+        adata = scp.datasets.isogenic_fibroblast_cell_line()
+
+        assert adata.shape == (19, 52995)
+        assert adata.obs_names.tolist() == [
+            "SKN2-4-G1",
+            "SKN2-6-G1",
+            "SKN2-7-G1",
+            "SKN2-7-G2",
+            "SKN2-8-G2",
+            "SKN2-1-G2",
+            "SKN2-2-G1",
+            "SKN2-2-G2",
+            "SKN2-1-G1",
+            "SKN2-9-G1",
+            "SKN2-3-G2",
+            "SKN2-3-G1",
+            "SKN2-4-G2",
+            "SKN2-10-G2",
+            "SKN2-5-G2",
+            "SKN2-9-G2",
+            "SKN2-6-G2",
+            "SKN2-8-G1",
+            "SKN2-5-G1",
+        ]
+        assert adata.obs["cell_cycle_gate"].value_counts(sort=False).to_dict() == {
+            "G1/0": 9,
+            "G2/M": 10,
+        }
+        assert adata.obs["sra_run"].tolist() == [
+            "SRR1663132",
+            "SRR1663134",
+            "SRR1663135",
+            "SRR1663140",
+            "SRR1663142",
+            "SRR1661707",
+            "SRR1661705",
+            "SRR1661708",
+            "SRR1661704",
+            "SRR1663137",
+            "SRR1661709",
+            "SRR1661706",
+            "SRR1661710",
+            "SRR1663144",
+            "SRR1663138",
+            "SRR1663143",
+            "SRR1663139",
+            "SRR1663136",
+            "SRR1663133",
+        ]
+        np.testing.assert_array_equal(
+            adata.obs["figure_column"], np.arange(1, 20, dtype=np.int16)
+        )
+        assert adata.obs.loc["SKN2-3-G1", "source_cell_id"] == "cell1"
+        assert adata.obs.loc["SKN2-3-G1", "coverage_depth"] == 92
+        assert adata.obs.loc["SKN2-10-G2", "source_cell_id"] == "cell19"
+        assert adata.obs.loc["SKN2-10-G2", "coverage_depth"] == 47
+        obs_payload = (
+            "\n".join(
+                f"{index}\t{row.source_cell_id}\t{row.source_sample_id}\t"
+                f"{row.cell_cycle_gate}\t{row.sra_run}\t{row.coverage_depth}\t"
+                f"{row.coverage_breadth:.4f}\t{row.on_target_fraction:.6f}\t"
+                f"{row.off_target_fraction:.6f}\t{row.mapped_reads}\t"
+                f"{row.duplicates}\t{row.duplicate_fraction:.7f}\t"
+                f"{row.source_table_order}\t{row.figure_column}\t{row.figure_label}"
+                for index, row in zip(
+                    adata.obs_names, adata.obs.itertuples(), strict=True
+                )
+            )
+            + "\n"
+        )
+        assert (
+            hashlib.sha256(obs_payload.encode()).hexdigest()
+            == "80ed75d96e8f65f63a000b7bb1b7cbc3d994dd2e3608a49b9244bbfe3c12e2ae"
+        )
+
+        values, counts = np.unique(adata.X, return_counts=True)
+        assert dict(zip(values, counts, strict=True)) == {
+            0: 26797,
+            1: 973715,
+            3: 6393,
+        }
[…]
