### 2026-07-19 93ba198
✨ Add Phyolin AML simulation collection
