+++ tests/test_datasets.py
@@ -1,6 +1,7 @@
 """Verify simulation, noise generation, and bundled dataset loaders."""
 
 import hashlib
+from pathlib import Path
 
 import numpy as np
 import pytest
@@ -300,6 +301,224 @@ class TestDatasets:
             "beta": 0.2521591,
             "gama": 0,
         }
+
+    def test_acute_myeloid_leukemia_many_collection(self):
+        """Verify all 120 independent Phyolin Figure 3 AML simulations."""
+        assert not hasattr(scp.datasets, "acute_lymphocytic_leukemia_many")
+        simulations = scp.datasets.acute_myeloid_leukemia_many()
+
+        assert len(simulations) == 120
+        assert list(simulations)[:3] == [
+            "AML-2_rep1",
+            "AML-2_rep2",
+            "AML-2_rep3",
+        ]
+        assert list(simulations)[-3:] == [
+            "AML-74_rep8",
+            "AML-74_rep9",
+            "AML-74_rep10",
+        ]
+        assert sum(adata.n_obs for adata in simulations.values()) == 872740
+        assert (
+            sum(adata.n_obs * adata.n_vars for adata in simulations.values()) == 3494950
+        )
+        assert sum(adata.n_vars for adata in simulations.values()) == 490
+        assert sum(np.sum(adata.X == 0) for adata in simulations.values()) == 1908984
+        assert sum(np.sum(adata.X == 1) for adata in simulations.values()) == 1585966
+        assert (
+            sum(np.sum(adata.layers["phyolin"] == 0) for adata in simulations.values())
+            == 1716567
+        )
+        assert (
+            sum(np.sum(adata.layers["phyolin"] == 1) for adata in simulations.values())
+            == 1778383
+        )
+        assert all(adata.X.dtype == np.int8 for adata in simulations.values())
+        assert all(
+            adata.layers["phyolin"].dtype == np.int8 for adata in simulations.values()
+        )
+
+        patient_shapes = {
+            "AML-2": (7931, 5),
+            "AML-8": (4675, 3),
+            "AML-10": (8729, 4),
+            "AML-33": (8120, 3),
+            "AML-47": (6497, 3),
+            "AML-58": (8170, 3),
+            "AML-53": (8013, 3),
+            "AML-62": (4027, 6),
+            "AML-63": (8347, 4),
+            "AML-67": (6024, 7),
+            "AML-69": (7462, 3),
+            "AML-74": (9279, 5),
+        }
+        for patient, expected_shape in patient_shapes.items():
+            patient_data = [
+                adata
+                for adata in simulations.values()
+                if adata.uns["simulation"]["patient_id"] == patient
+            ]
+            assert len(patient_data) == 10
+            assert {adata.shape for adata in patient_data} == {expected_shape}
+
+        provenance = simulations["AML-2_rep1"].uns["provenance"]
+        assert provenance["collection"] == {
+            "branched_instances": 60,
+            "correct_topology_classifications": 120,
+            "distinct_mutation_labels": 37,
+            "input_ones": 1585966,
+            "input_zeros": 1908984,
+            "linear_instances": 60,
+            "n_instances": 120,
+            "n_patients": 12,
+            "output_ones": 1778383,
+            "output_zeros": 1716567,
+            "phyolin_classification_threshold": 0.05,
+            "phyolin_flips_0_to_1": 192417,
+            "replicates_per_patient": 10,
+            "runtime_limit_reached": 24,
+            "runtime_limit_seconds": 500,
+            "simulated_false_negative_rate": 0.05,
+            "simulated_false_negatives": 127275,
+            "total_cells": 872740,
[…]
