+++ src/scphylo/datasets/__init__.py
@@ -9,6 +9,7 @@ from scphylo.datasets._datasets import (
     acute_lymphocytic_leukemia6,
     acute_myeloid_leukemia1,
     acute_myeloid_leukemia2,
+    acute_myeloid_leukemia_many,
     colorectal1,
     colorectal2,
     colorectal3,
@@ -40,6 +41,7 @@ __all__ = (
     acute_lymphocytic_leukemia6,
     acute_myeloid_leukemia1,
     acute_myeloid_leukemia2,
+    acute_myeloid_leukemia_many,
     colorectal1,
     colorectal2,
     colorectal3,
