+++ src/scphylo/datasets/_datasets.py
@@ -1,3 +1,10 @@
+import copy
+import json
+
+import anndata as ad
+import numpy as np
+import pandas as pd
+
 import scphylo as scp
 
 
@@ -730,20 +737,135 @@ def oligodendroglioma_idh_mutated_tumor():
     return adata
 
 
-def acute_lymphocytic_leukemia_many():
-    """Human Acute Lymphocytic Leukemia datasets.
+_AML_MANY_ASSET = "scphylo.datasets/real/acute_myeloid_leukemia_many.npz"
+_AML_MANY_ASSET_SHA256 = (
+    "edd2caea3ce7750a3c0f9f7ec1ba4dfaedc9ed2c52f32b958551f12d17dcafd5"
+)
 
-    This dataset was introduced in :cite:`Morita_2020` and was used in:
 
-    * :cite:`Phyolin` Figure 3 and Table 1.
+def _phyolin_aml_simulation(archive, metadata, instance):
+    """Construct one Phyolin AML simulation from the compressed collection."""
+    instance_id = instance["instance_id"]
+    matrix = np.array(archive[f"input__{instance_id}"], dtype=np.int8, copy=True)
+    phyolin = np.array(archive[f"phyolin__{instance_id}"], dtype=np.int8, copy=True)
+    n_cells, n_mutations = matrix.shape
+
+    obs_names = pd.Index(
+        (f"source_cell_{row:05d}" for row in range(1, n_cells + 1)),
+        name="cell",
+    )
+    categorical_codes = np.zeros(n_cells, dtype=np.int8)
+    obs = pd.DataFrame(index=obs_names)
+    obs["source_row"] = np.arange(1, n_cells + 1, dtype=np.int32)
+    obs["instance_id"] = pd.Categorical.from_codes(
+        categorical_codes, categories=[instance_id]
+    )
+    obs["patient_id"] = pd.Categorical.from_codes(
+        categorical_codes, categories=[instance["patient_id"]]
+    )
+    obs["replicate"] = np.full(n_cells, instance["replicate"], dtype=np.int8)
+    obs["published_topology"] = pd.Categorical.from_codes(
+        categorical_codes, categories=[instance["published_topology"]]
+    )
+
+    labels = instance["mutation_labels"]
+    var = pd.DataFrame(
+        {
+            "source_label": labels,
+            "source_column": np.arange(1, n_mutations + 1, dtype=np.int16),
+            "gene_as_reported": [
+                label.split("_", 1)[0].split(".", 1)[0].removesuffix("-ITD")
+                for label in labels
+            ],
+        },
+        index=pd.Index(labels, name="mutation"),
+    )
+
+    adata = ad.AnnData(X=matrix, obs=obs, var=var)
+    adata.layers["phyolin"] = phyolin
+    adata.uns["simulation"] = copy.deepcopy(instance)
+    adata.uns["simulation"]["simulated_false_negative_rate"] = metadata["collection"][
+        "simulated_false_negative_rate"
+    ]
+    adata.uns["provenance"] = copy.deepcopy(metadata["provenance"])
+    adata.uns["provenance"]["collection"] = copy.deepcopy(metadata["collection"])
+    adata.uns["provenance"]["asset"] = {
+        "package_resource": _AML_MANY_ASSET,
+        "sha256": _AML_MANY_ASSET_SHA256,
+        "schema_version": metadata["schema_version"],
+    }
+    return adata
+
+
+def acute_myeloid_leukemia_many(instance=None):
+    """Load the Phyolin simulations approximating an AML cohort.
+
+    :cite:`Phyolin` Figure 3 and Table 1 used 12 acute myeloid leukemia
+    (AML) topologies derived from :cite:`Morita_2020`, with 10 independently
[…]
