### 2026-07-19 bbb171d
✅ Add unit tests for new features
