+++ tests/test_tl_fast_coverage.py
@@ -1,0 +1,265 @@
+"""Cover small tree-tool branches without invoking expensive solver backends."""
+
+from decimal import Decimal
+from unittest.mock import Mock
+
+import networkx as nx
+import numpy as np
+import pandas as pd
+import pytest
+
+import scphylo as scp
+from scphylo.tl.consensus import _consensus as consensus_module
+from scphylo.tl.partition_function import _clt_sampler, _partition_function, _pf
+from scphylo.tl.score import _others, _ours
+
+
+def _matrix(values, *, cells=("c1", "c2", "c3"), muts=("a", "b")):
+    """Build a small labeled genotype matrix."""
+    return pd.DataFrame(values, index=cells, columns=muts)
+
+
+@pytest.mark.parametrize(
+    ("score", "ground", "solution"),
+    [
+        (
+            _ours.gs,
+            _matrix([[1], [0]], cells=("c1", "c2"), muts=("a",)),
+            _matrix([[1], [0]], cells=("c1", "c2"), muts=("b",)),
+        ),
+        (
+            _ours.gs,
+            _matrix([[1]], cells=("c1",), muts=("a",)),
+            _matrix([[1]], cells=("c2",), muts=("a",)),
+        ),
+        *[
+            (
+                score,
+                _matrix([[1], [0]], cells=("c1", "c2"), muts=("a",)),
+                _matrix([[1], [0]], cells=("c1", "c2"), muts=("b",)),
+            )
+            for score in (
+                _ours.ad,
+                _ours.dl,
+                _ours.cc,
+                _ours.mltd,
+                _ours.tpted,
+                _others.mp3,
+                _others.caset,
+                _others.disc,
+            )
+        ],
+        (
+            _others.rf,
+            _matrix([[1]], cells=("c1",), muts=("a",)),
+            _matrix([[1]], cells=("c2",), muts=("a",)),
+        ),
+    ],
+)
+def test_scores_reject_inputs_without_shared_labels(score, ground, solution):
+    """Reject trees that have no mutations or cells available for comparison."""
+    with pytest.raises(RuntimeError):
+        score(ground, solution)
+
+
+@pytest.mark.parametrize(
+    ("ground", "solution"),
+    [
+        (
+            _matrix([[1, 1], [1, 0], [0, 0]]),
+            _matrix([[1, 0], [0, 1], [0, 0]]),
+        ),
+        (
+            _matrix([[1, 1], [0, 1], [0, 0]]),
+            _matrix([[1, 1], [1, 0], [0, 0]]),
+        ),
+    ],
+)
+def test_ad_detects_each_reversed_ancestor_case(ground, solution):
+    """Count disjoint and direction-reversed ancestor relationships as errors."""
+    assert _ours.ad(ground, solution) == 0
+
+
+def test_ad_rejects_a_tree_without_ancestor_pairs():
+    """Reject a comparison for which the ground tree has no ancestor pair."""
+    matrix = _matrix([[1], [0], [0]], muts=("a",))
+
+    with pytest.raises(RuntimeError):
+        _ours.ad(matrix, matrix)
+
+
+def test_unimplemented_scores_return_none():
+    """Document the current placeholder behavior of Bourque and PCSS scores."""
+    matrix = _matrix([[1, 0], [0, 1], [0, 0]])
+
+    assert _others.bourque(matrix, matrix) is None
+    assert _others.pcss(matrix, matrix) is None
+
+
+def test_mltd_reports_a_core_failure(monkeypatch):
+    """Turn a failed native MLTD call into the package's standard runtime error."""
+    matrix = _matrix([[1, 1], [1, 0], [0, 0]])
[…]
