+++ src/scphylo/__init__.py
@@ -6,5 +6,5 @@ from scphylo import logging as logg
 __author__ = ", ".join(["Farid Rashidi"])
 __maintainer__ = ", ".join(["Farid Rashidi"])
 __email__ = ", ".join(["farid.rsh@gmail.com"])
-__version__ = "0.0.5"
+__version__ = "0.1.0"
 __all__ = (datasets, io, logg, pl, pp, settings, tl, ul)
