+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "scphylo-tools"
-version = "0.0.5"
+version = "0.1.0"
 description = "A python toolkit for single-cell tumor phylogenetic analysis"
 requires-python = ">=3.11"
 dependencies = [
@@ -107,7 +107,7 @@ where = ["src"]
 build = "cp31{1,2,3}-*"
 
 [tool.bumpversion]
-current_version = "0.0.5"
+current_version = "0.1.0"
 commit = true
 message = "🔖 Release {new_version}"
 tag = true
