### 2026-08-25 #10589 @willgearty
Everything looks great to me. I'm happy to recommend this submission for acceptance!
