### 2026-08-25 #10589 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
