### 2026-09-01 #10589 @Kevin-Mattheus-Moerman
@faridrashidi As AEiC for JOSS I will now help to process this submission for acceptance in JOSS. Below are some final checks:

- [x] Copyedit pass/checked for typos

## Comment / recommendations (not required to take action)
* Your project image in the README is hard to appreciate when folks use GitHub dark mode, e.g. for me it looked like this:

As such I would recommend you add a white background to your image.
