### 2026-09-01 #10589 @faridrashidi
@Kevin-Mattheus-Moerman, Thank you for reviewing! I added separate versions of the project image for light and dark themes and updated the README to include the citation reference and CITATION.cff.
