### 2026-09-01 #10589 @Kevin-Mattheus-Moerman
@faridrashidi congratulations on this JOSS paper!

@willgearty thanks for editing!

And a special thank you to the reviewers: @rachelicr, @shrishar !!!!!
