### 2026-09-01 #10589 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
